Rename checkConnection
The previous method name could give the impression that the connection
actually would be checked when the method was called.

What the method actually does it just to report the flag that is set when the
initial connection has been established and the client is ready to receive data
from the broker/server. That should hopefully be more clear after the change of name.

Old name kept on "top level" to not introduce any backward incompatible changes.

Fixes https://github.com/eclipse/kuksa.val/issues/523